fix: precision on resetUnstakeLockPenalty increased

+++ contracts/Core/parameters/Governance.sol
@@ -124,7 +124,7 @@ contract Governance is Initializable, ACL, Constants {
      * @dev can be called only by the the address that has the governer role
      * @param _extendLockPenalty updated value to be set for extendLockPenalty
      */
-    function setResetUnstakeLockPenalty(uint8 _extendLockPenalty) external initialized onlyRole(GOVERNER_ROLE) {
+    function setResetUnstakeLockPenalty(uint32 _extendLockPenalty) external initialized onlyRole(GOVERNER_ROLE) {
         emit ParameterChanged(msg.sender, "extendLockPenalty", _extendLockPenalty, block.timestamp);
         stakeManagerParams.setResetUnstakeLockPenalty(_extendLockPenalty);
     }
